Move tests to BridgeTests class
